Add player upgrade level
- Assign value from player upgrade data and gameplay data.

+++ Assets/NumGates/Scripts/Scriptables/GameplayValueData.cs
@@ -8,7 +8,7 @@ namespace NumGates
     public class GameplayValueData : ScriptableObject
     {
         public int[] healthValue;
-        public float[] shieldTimerValue;
+        public int[] shieldTimerValue;
         public int[] pureSoulValue;
         public int[] colorSoulValue;
         public int[] cryptoValue;
